+++ erdjs/src/networkConfig.ts
@@ -1,12 +1,6 @@
 import BigNumber from "bignumber.js";
 import { IProvider } from "./interface";
-import {
-    GasPrice,
-    GasLimit,
-    TransactionVersion,
-    ChainID,
-    GasPriceModifier,
-} from "./networkParams";
+import { GasPrice, GasLimit, TransactionVersion, ChainID, GasPriceModifier } from "./networkParams";
 
 /**
  * An object holding Network configuration parameters.
@@ -23,6 +17,14 @@ export class NetworkConfig {
      * The gas required by the Network to process a byte of the {@link TransactionPayload}.
      */
     public GasPerDataByte: number;
+    /**
+     * The round duration.
+     */
+    public RoundDuration: number;
+    /**
+     * The number of rounds per epoch.
+     */
+    public RoundsPerEpoch: number;
 
     /**
      * The Top Up Factor for APR calculation
@@ -58,6 +60,8 @@ export class NetworkConfig {
         this.ChainID = new ChainID("T");
         this.GasPerDataByte = 1500;
         this.TopUpFactor = 0;
+        this.RoundDuration = 0;
+        this.RoundsPerEpoch = 0;
         this.TopUpRewardsGradientPoint = new BigNumber(0);
         this.MinGasLimit = new GasLimit(50000);
         this.MinGasPrice = new GasPrice(1000000000);
@@ -94,17 +98,13 @@ export class NetworkConfig {
         networkConfig.ChainID = new ChainID(payload["erd_chain_id"]);
         networkConfig.GasPerDataByte = Number(payload["erd_gas_per_data_byte"]);
         networkConfig.TopUpFactor = Number(payload["erd_top_up_factor"]);
-        networkConfig.TopUpRewardsGradientPoint = new BigNumber(
-            payload["erd_rewards_top_up_gradient_point"]
-        );
+        networkConfig.RoundDuration = Number(payload["erd_round_duration"]);
+        networkConfig.RoundsPerEpoch = Number(payload["erd_rounds_per_epoch"]);
+        networkConfig.TopUpRewardsGradientPoint = new BigNumber(payload["erd_rewards_top_up_gradient_point"]);
         networkConfig.MinGasLimit = new GasLimit(payload["erd_min_gas_limit"]);
         networkConfig.MinGasPrice = new GasPrice(payload["erd_min_gas_price"]);
-        networkConfig.MinTransactionVersion = new TransactionVersion(
-            payload["erd_min_transaction_version"]
-        );
-        networkConfig.GasPriceModifier = new GasPriceModifier(
-            payload["erd_gas_price_modifier"]
-        );
+        networkConfig.MinTransactionVersion = new TransactionVersion(payload["erd_min_transaction_version"]);
+        networkConfig.GasPriceModifier = new GasPriceModifier(payload["erd_gas_price_modifier"]);
 
         return networkConfig;
     }